w9 (forgot to make a build whoops)
Hey Brendan, I'm assuming you're the one marking this. Sorry for making it a pain in the ass having two separate repositories with pretty awful naming schemes, however starting from scratch did genuinely allow me to properly understand C# as well as the conventions of coding.

In addition to this, as there's no written report component with this assessment I wanted to mention that most of my independent re-doing was following the YT video 'How to make Asteroids in Unity' by Zigurous (https://www.youtube.com/watch?v=cIeWhztKyAg), I didn't follow it completely, as the video doesn't include various UI features as well as the players' movement code we had used in class as well as other aspects, but since I did use it as a resource I thought it would be better to mention then not.

thank you for you time